backend/pyproject.toml
[project]
name = "easy-painter-backend"
version = "0.1.0"
description = "FastAPI backend for the Easy Painter public image generation site."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.4.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "minio>=7.2.15",
    "orjson>=3.10.16",
    "psycopg[binary]>=3.2.6",
    "pydantic-settings>=2.8.1",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.40",
    "uvicorn[standard]>=0.34.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
